+++ frontend/src/lib/routes.ts
@@ -2,6 +2,7 @@ import { FiBook, FiUser, FiBriefcase, FiEdit } from 'react-icons/fi';
 import { IconType } from 'react-icons';
 import { VscVersions } from 'react-icons/vsc';
 import { FiLogOut } from 'react-icons/fi';
+import { MdOutlineSupportAgent } from 'react-icons/md';
 
 // ルートの中身を定義
 export interface Route {
@@ -15,6 +16,7 @@ export const NAV_LINKS: Route[] = [
   { label: 'Wiki', href: '/wiki', icon: FiBook },
   { label: 'Profile', href: '/profile', icon: FiUser },
   { label: 'Portfolio', href: '/portfolio', icon: FiBriefcase },
+  { label: 'Chat', href: '/chat', icon: MdOutlineSupportAgent },
   { label: 'Log', href: '/changelog', icon: VscVersions },
 ];
 
@@ -24,6 +26,7 @@ export const NAV_LINKS_WITH_AUTH: Route[] = [
   { label: 'Admin', href: '/admin', icon: FiEdit },
   { label: 'Profile', href: '/profile', icon: FiUser },
   { label: 'Portfolio', href: '/portfolio', icon: FiBriefcase },
+  { label: 'Chat', href: '/chat', icon: MdOutlineSupportAgent },
   { label: 'Log', href: '/changelog', icon: VscVersions },
 ];